refactor: properly integrate upgrade API
