+++ packages/openapi-parser/src/utils/index.ts
@@ -2,5 +2,8 @@ export * from './normalize'
 export * from './resolve'
 export * from './toJson'
 export * from './toYaml'
+export * from './traverse'
 export * from './upgrade'
+export * from './upgradeFromThreeToThreeOne'
+export * from './upgradeFromTwoToThree'
 export * from './validate'
