+++ packages/openapi-parser/src/utils/upgradeFromTwoToThree.test.ts
@@ -1,0 +1,18 @@
+import { describe, expect, it } from 'vitest'
+
+import { upgradeFromTwoToThree } from './upgradeFromTwoToThree'
+
+describe.skip('version', () => {
+  it('doesn’t modify Swagger 2.0 files', async () => {
+    const result = upgradeFromTwoToThree({
+      swagger: '2.0',
+      info: {
+        title: 'Hello World',
+        version: '1.0.0',
+      },
+      paths: {},
+    })
+
+    expect(result.openapi).toBe('3.0.0')
+  })
+})
