+++ packages/openapi-parser/src/utils/upgradeFromTwoToThree.ts
@@ -1,0 +1,9 @@
+/**
+ * Upgrade Swagger 2.0 to OpenAPI 3.0
+ *
+ * https://swagger.io/blog/news/whats-new-in-openapi-3-0/
+ */
+export function upgradeFromTwoToThree(specification: Record<string, any>) {
+  // TODO: Implement
+  return specification
+}
